[MDH-202] refactor : 웨이팅 알람 패키지 이동
